Changes to file names and seed working with multiple csv files
